+++ RayTracingProject/SourceAndHeadersOnly/point.h
@@ -4,6 +4,7 @@
 #include <unistd.h>
 #include <math.h>
 #include <SDL/SDL.h>
+#include "wall.h"
 
 #ifndef POINT_H_INCLUDED
 #define POINT_H_INCLUDED
@@ -16,8 +17,8 @@ typedef struct POINT POINT;
 
 struct POINT {
     /* Donne les coordonnées d'un point */
-    int x;
-    int y;
+    float x;
+    float y;
 };
 
 void pointImage(POINT *pointImage, POINT copier, POINT comparer, WALL wall);
